Initial docker files
I don't know if I'm convinced that docker is the right way to do this.
It seems not very clean and not very applicable to our prod system.  I
say that because best practices suggest that you shouldn't keep dbs in
containers, so we won't on prod.  Plus, we were planning on having a
staging db for testing and whatnot, so we may as well just point our
tests there and make them clean up after themselves/be sufficiently
random.

As of right now, docker seems like a slow, bloated, confusing way of
handling recreatable tests.  Because of that I think I'll be abandoning
it (for now).